+++ .github/workflows/fuzzing.yml
@@ -33,7 +33,7 @@ jobs:
         - name: fuzz telio crypto decrypt_response
           working-directory: crates/telio-crypto
           run: cargo fuzz run decrypt_response -- -max_total_time=180
-  telio-wg-fuzzing:
+  telio-pq-fuzzing:
       runs-on: ubuntu-22.04
       steps:
         - uses: actions/checkout@c85c95e3d7251135ab7dc9ce3241c5835cc595a9 # v3.5.3
@@ -42,6 +42,6 @@ jobs:
             toolchain: nightly-2023-06-01
             override: true
         - run: cargo install cargo-fuzz --locked --version 0.11.2
-        - name: fuzz telio wg parse_get_packet
-          working-directory: crates/telio-wg
+        - name: fuzz telio pq parse_get_packet
+          working-directory: crates/telio-pq
           run: cargo fuzz run parse_get_packet -- -max_total_time=180
