+++ crates/telio-pq/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "telio-pq"
+version = "0.1.0"
+edition = "2021"
+license = "GPL-3.0-only"
+repository = "https://github.com/NordSecurity/libtelio"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+fuzzing = []
+
+[dependencies]
+# pqcrypto version is fixed, because the newer version implements incompatible kyber kem according to draft specs
+pqcrypto-kyber = { version = "=0.7.6", default-features = false, features = ["std"] }
+pqcrypto-traits = "0.3.5"
+
+telio-crypto.workspace = true
+telio-model.workspace = true
+telio-sockets.workspace = true
+telio-task.workspace = true
+telio-utils.workspace = true
+
+boringtun.workspace = true
+pnet_packet.workspace = true
+rand.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+
+[dev-dependencies]
+base64.workspace = true
