+++ crates/telio-pq/fuzz/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "telio-wg-fuzz"
+name = "telio-pq-fuzz"
 version = "0.0.0"
 publish = false
 edition = "2021"
@@ -12,7 +12,7 @@ cargo-fuzz = true
 [dependencies]
 libfuzzer-sys = "0.4"
 
-[dependencies.telio-wg]
+[dependencies.telio-pq]
 path = ".."
 features = ["fuzzing"]
