+++ crates/telio-pq/fuzz/fuzz_targets/parse_get_packet.rs
@@ -3,5 +3,5 @@
 use libfuzzer_sys::fuzz_target;
 
 fuzz_target!(|data: &[u8]| {
-    let _result = telio_wg::pq::parse_get_response_fuzz(data);
+    let _result = telio_pq::fuzz::parse_get_response(data);
 });
