+++ crates/telio-wg/Cargo.toml
@@ -8,12 +8,8 @@ publish = false
 
 [features]
 test-adapter = []
-fuzzing = []
 
 [dependencies]
-# pqcrypto version is fixed, because the newer version implements incompatible kyber kem according to draft specs
-pqcrypto-kyber = { version = "=0.7.6", default-features = false, features = ["std"] }
-pqcrypto-traits = "0.3.5"
 slog-stdlog = "4.1.0"
 wireguard-uapi = { version = "2.0.4", features = ["xplatform"]}
 
@@ -26,11 +22,10 @@ lazy_static.workspace = true
 libc.workspace = true
 tracing.workspace = true
 mockall = { workspace = true, optional = true }
-pnet_packet.workspace = true
+rand.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 slog.workspace = true
-rand.workspace = true
 thiserror.workspace = true
 tokio = { workspace = true, features = ["full"] }
 
@@ -41,7 +36,6 @@ telio-task.workspace = true
 telio-utils.workspace = true
 
 [dev-dependencies]
-base64.workspace = true
 mockall.workspace = true
 ntest.workspace = true
 pretty_assertions.workspace = true
