+++ crates/telio-wg/src/lib.rs
@@ -8,8 +8,6 @@ pub(crate) mod adapter;
 pub(crate) mod wg;
 pub(crate) mod windows;
 
-/// Post quantum primitives
-pub mod pq;
 pub mod uapi;
 
 pub use crate::{
